downloads_mp3/requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
ipython==7.29.0
Werkzeug==2.0.2
spotipy==2.23.0
youtube_dl==2021.12.17
youtube_search==2.1.2
yt_dlp==2024.3.10
ffprobe==0.5
ffmpeg==1.4
pandas == 2.2.1
black==24.3.0
python-dotenv==1.0.1